Changing data type of OwnerType from tinyint to int
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C52608

+++ TOneV2/Code/TOneV2/TOne.WhS.CodePreparation.Data.SQL/SaleZoneRPPreviewDataManager.cs
@@ -83,7 +83,7 @@ namespace TOne.WhS.CodePreparation.Data.SQL
             ZoneRoutingProductPreview zoneRoutingProductPreview = new ZoneRoutingProductPreview
             {
                 ZoneName = GetReaderValue<string>(reader, "ZoneName"),
-                OwnerType = (SalePriceListOwnerType)GetReaderValue<byte>(reader, "OwnerType"),
+                OwnerType = (SalePriceListOwnerType)GetReaderValue<int>(reader, "OwnerType"),
                 OwnerId = (int)reader["OwnerID"],
                 RoutingProductId = (int)reader["RoutingProductID"],
                 BED = (DateTime)reader["BED"],
